+++ source/game_base/ascii_view.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 
-from tictactoe.tictactoe import Token
+from game_base.board import Token
 
 
 class ASCII_View():
@@ -11,22 +11,31 @@ class ASCII_View():
         self._grid = grid
         self._messages = list()
         self.current_player = None
+        self._history = list()
 
     def welcome(self, title, version):
         print('{} {}'.format(title, version))
 
     def display(self):
+        self._print_history()
         self._print_messages()
         print('-----')
         self._print_grid(self._grid)
         self._print_player(self.current_player)
 
-    def ask_input(self):
-        i = input("Enter (x,y): ")
-        return int(i[0]), int(i[1])
+    def ask_input(self, n):
+        if n == 1:
+            i = input("Enter (x): ")
+            return int(i)
+        else:
+            i = input("Enter (x,y): ")
+            return int(i[0]), int(i[1])
 
-    def message(self, msg):
+    def add_message(self, msg):
         self._messages.append(msg)
+
+    def set_history(self, h):
+        self._history = h
 
     def _print_grid(self, grid):
 
@@ -47,9 +56,9 @@ class ASCII_View():
     def _token_str(self, token):
         if token is None:
             return '-'
-        elif token in [Token.CROSS]:
+        elif token in [Token.A]:
             return 'X'
-        elif token in [Token.CIRCLE]:
+        elif token in [Token.B]:
             return 'O'
         else:
             assert False
@@ -59,3 +68,10 @@ class ASCII_View():
             print(msg)
 
         self._messages.clear()
+
+    def _print_history(self):
+        if self._history:
+            print('History:', end='')
+            for m in self._history:
+                print('{},'.format(m), end='')
+            print()