+++ test/game/test_tictactoe.py
@@ -6,7 +6,7 @@ import sys
 import os
 sys.path.insert(0, os.path.join(sys.path[0], 'source'))
 
-from tictactoe.tictactoe import (TicTacToe, Token)
+from game.tictactoe import (TicTacToe, Token)
 from game_base.board import (Point)
 
 
@@ -31,6 +31,12 @@ class TestTicTacToe(unittest.TestCase):
         # Cell not free, do not change current player
         self.assertFalse(game.play(Point(1, 0)))
         self.assertEqual(game._current_player, game._p1)
+
+    def test_compute_ending(self):
+        game = TicTacToe()
+        self.assertEqual(game._board.cell_used_count, 0)
+        game.play(Point(0, 0))
+        self.assertEqual(game._board.cell_used_count, 1)
 
     def test_win_horizontal_player1(self):
         game = TicTacToe()
@@ -95,7 +101,7 @@ class TestTicTacToe(unittest.TestCase):
         self.assertTrue(game.is_over)
         self.assertEqual(game.winner, game._p2)
 
-    def test_win_diag_player1(self):
+    def test_win_diag_down_player1(self):
         game = TicTacToe()
         self.assertTrue(game.play(Point(0, 0)))
         self.assertTrue(game.play(Point(1, 0)))
@@ -106,7 +112,7 @@ class TestTicTacToe(unittest.TestCase):
         self.assertTrue(game.is_over)
         self.assertEqual(game.winner, game._p1)
 
-    def test_win_diag_player2(self):
+    def test_win_diag_up_player2(self):
         game = TicTacToe()
         self.assertTrue(game.play(Point(0, 0)))
         self.assertTrue(game.play(Point(0, 2)))
@@ -117,6 +123,31 @@ class TestTicTacToe(unittest.TestCase):
 
         self.assertTrue(game.is_over)
         self.assertEqual(game.winner, game._p2)
+
+    def test_win_diag_up_p1(self):
+        '''
+          0 1 2
+        0|X|O|X|
+        1|O|X|O|
+        2|X|-|-|
+        '''
+        game = TicTacToe()
+        self.assertTrue(game.play(Point(0, 0)))
+        self.assertTrue(game.play(Point(1, 0)))
+        self.assertTrue(game.play(Point(2, 0)))
+        self.assertTrue(game.play(Point(0, 1)))
+        self.assertTrue(game.play(Point(1, 1)))
+        self.assertTrue(game.play(Point(2, 1)))
+        self.assertTrue(game.play(Point(0, 2)))
+
+        self.assertTrue(game.is_over)
+        self.assertEqual(game.winner, game._p1)
+
+    def test_generate_moves(self):
+        game = TicTacToe()
+        moves = game.generate_moves()
+        self.assertEqual(moves, [Point(0, 0), Point(1, 0), Point(2, 0), Point(
+            0, 1), Point(1, 1), Point(2, 1), Point(0, 2), Point(1, 2), Point(2, 2)])
 
 
 if __name__ == '__main__':